libtrellis/CMakeLists.txt: Use GNUInstallDirs to determine install paths
Use the GNUInstallDirs include provided by CMake to auto detect target
install paths for the target system. This allows for building natively
on the same host as well as enabling a distro to override the paths or
for cross compiling to override the paths for the target distro/system.

This removes the need to manually detect the 'lib64' directory, and also
fixes the issues with using 'FIND_LIBRARY_USE_LIB64_PATHS' which is
always true on targets other than debian (even if those targets do not
use the 'lib64' path by default).

Additionally this change calculates the relative path between bindir and
libdir in order to populate the rpath information into the built
binaries for supporting portable execution.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>