libtrellis: Add support for portable execution with trellis database
In order to execute the ecpmulti, ecppack and ecpunpack programs they
must be able to access the trellis database files. These files are
installed alongside the tools in the data directory (normally
/usr/share). To enable portable execution of these programs such that
they can access the database relative to their executables path the
programs must be built with the relative path embedded and also be able
to access the executables path at run time.

Since libtrellis already depends on Boost it makes sense to use the
Boost dll module. This module provides the 'program_location' function
which can be used with the 'boost::filesystem::path' to build the
absolute path to the database at run time. The only downside with the
use of the Boost dll module is that a unused component of the included
source relies on certain dynamic linker functions which requires linking
to libdl on some platforms. This however is easily handled by the
CMAKE_DL_LIBS variable which expands to the correct linker arg on
targets that require linking with the dynamic linker.

The Boost dll module was only added in 1.61. In order to support older
versions of Boost such as that of Ubuntu 16.04 (1.58), a local
implementation of 'program_location' is added which is based on the
'proc_self_dirname' implementation from yosys. This version is chosen
due to its use in a number of tools similar to libtrellis (e.g.
icestorm) but also due to its ISC license which matches libtrellis. This
implementation is conditionally used if the Boost dll module is
unavailable.

The cmake configuration is also changed to generate a relative path from
datadir to bindir, this path is then provided during the build as
TRELLIS_RPATH_DATADIR.

Note: ecppll does not currently use the trellis database

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>