libtrellis: version: fix linking on MSVC
MSVC has a concept of "compilation units".  It doesn't appear possible
to have an `extern` symbol without explicitly exporting it.  It is
unclear exactly why this is.

However, performing `#include "version.hpp"` appears to fix the problem,
as both compilation units now refer to the exact same symbol.

This fixes the build on MSVC.

Signed-off-by: Sean Cross <sean@xobs.io>