Add ecppack --xsvf option (enabling XSVF output).
The produced XSVF makes use of the Lattice LCOUNT/LDELAY/LSDR extensions.

There are two main advantages to XSVF over SVF: 1) potentially smaller
output files (when the increased density of binary encoding outweighs
the overhead of mandatory TDOs and masks in XSVF); and 2) the LSDR
instruction allows polling until the TAP is ready to accept the next
command.  This is particularly valuable when writing to SPI flash, since
the required delay can vary widely depending on the flash device, and
without repeated polling we are forced to impose a conservative delay
before one check for completion.

Signed-off-by: Gary Wong <gtw@gnu.org>